pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "simple_agent_tester"
version = "0.3.0+ollama"
description = "A zero-code agent tester for local Ollama models."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT License" }
authors = [
    { name = "Ria Chaudhuri", email = "ria-xxxxxxxxx@gmail.com" },
]

# --- Your Package Dependencies ---
dependencies = [
    "ollama >= 0.2.0",   # For running the agent
    "PyYAML",            # For reading the config.yaml
    "typer[all]",        # For creating the beautiful CLI
    "rich",              # For beautiful terminal printing
    "importlib-resources" # To find the template file
]

# --- This creates the CLI command ---
[project.scripts]
agent-tester = "simple_agent_tester.cli:app"